refactor: split off db with migration
The entity module requires common, common requires migration due to the
DB setup functions.

However, if we want to use the entities in the migration as well, then
that would lead to a circular dependency situation.

Moving the setup methods into a special "db" modules solves this issue.
