+++ server/Cargo.toml
@@ -11,6 +11,7 @@ graphql = ["trustify-module-graphql"]
 [dependencies]
 trustify-auth = { workspace = true }
 trustify-common = { workspace = true }
+trustify-db = { workspace = true }
 trustify-infrastructure = { workspace = true }
 trustify-module-analysis = { workspace = true }
 trustify-module-fundamental = { workspace = true }
